Enabled multiplication optimizations; SealBFVBackend supports only float now

+++ test/test_add.in.cpp
@@ -38,8 +38,8 @@ TEST(${BACKEND_NAME}, add_2_3)
     auto backend = runtime::Backend::create("${BACKEND_REGISTERED_NAME}");
 
     Shape shape{2, 3};
-    auto a = make_shared<op::Parameter>(element::i64, shape);
-    auto b = make_shared<op::Parameter>(element::i64, shape);
+    auto a = make_shared<op::Parameter>(element::f32, shape);
+    auto b = make_shared<op::Parameter>(element::f32, shape);
     auto t = make_shared<op::Add>(a, b);
     auto f = make_shared<Function>(t, op::ParameterVector{a, b});
 
@@ -55,11 +55,11 @@ TEST(${BACKEND_NAME}, add_2_3)
         auto t_b = inputs[1];
         auto t_result = results[0];
 
-        copy_he_data(t_a, test::NDArray<int64_t, 2>({{1, 2, 3}, {4, 5, 6}}).get_vector(), backend);
-        copy_he_data(t_b, test::NDArray<int64_t, 2>({{7, 8, 9}, {10, 11, 12}}).get_vector(), backend);
+        copy_he_data(t_a, test::NDArray<float, 2>({{1, 2, 3}, {4, 5, 6}}).get_vector(), backend);
+        copy_he_data(t_b, test::NDArray<float, 2>({{7, 8, 9}, {10, 11, 12}}).get_vector(), backend);
         backend->call(f, {t_result}, {t_a, t_b});
-        EXPECT_EQ(read_he_vector<int64_t>(t_result, backend),
-                  (test::NDArray<int64_t, 2>({{8, 10, 12}, {14, 16, 18}})).get_vector());
+        EXPECT_EQ(read_he_vector<float>(t_result, backend),
+                  (test::NDArray<float, 2>({{8, 10, 12}, {14, 16, 18}})).get_vector());
     }
 }
 
@@ -68,8 +68,8 @@ TEST(${BACKEND_NAME}, add_zero_2_3)
     auto backend = runtime::Backend::create("${BACKEND_REGISTERED_NAME}");
 
     Shape shape{2, 3};
-    auto a = make_shared<op::Parameter>(element::i64, shape);
-    auto b = make_shared<op::Parameter>(element::i64, shape);
+    auto a = make_shared<op::Parameter>(element::f32, shape);
+    auto b = make_shared<op::Parameter>(element::f32, shape);
     auto t = make_shared<op::Add>(a, b);
     auto f = make_shared<Function>(t, op::ParameterVector{a, b});
 
@@ -85,10 +85,10 @@ TEST(${BACKEND_NAME}, add_zero_2_3)
         auto t_b = inputs[1];
         auto t_result = results[0];
 
-        copy_he_data(t_a, test::NDArray<int64_t, 2>({{1, 2, 3}, {4, 5, 6}}).get_vector(), backend);
-        copy_he_data(t_b, test::NDArray<int64_t, 2>({{0, 0, 0}, {0, 0, 0}}).get_vector(), backend);
+        copy_he_data(t_a, test::NDArray<float, 2>({{1, 2, 3}, {4, 5, 6}}).get_vector(), backend);
+        copy_he_data(t_b, test::NDArray<float, 2>({{0, 0, 0}, {0, 0, 0}}).get_vector(), backend);
         backend->call(f, {t_result}, {t_a, t_b});
-        EXPECT_EQ(read_he_vector<int64_t>(t_result, backend),
-                  (test::NDArray<int64_t, 2>({{1, 2, 3}, {4, 5, 6}})).get_vector());
+        EXPECT_EQ(read_he_vector<float>(t_result, backend),
+                  (test::NDArray<float, 2>({{1, 2, 3}, {4, 5, 6}})).get_vector());
     }
 }
